requirements.txt
docopt==0.6.2
Pillow
unidecode